+++ phoss-smp-backend/src/test/java/com/helger/phoss/smp/mock/MockSMPRedirectManager.java
@@ -46,7 +46,7 @@ final class MockSMPRedirectManager implements ISMPRedirectManager
     throw new UnsupportedOperationException ();
   }
 
-  public int getSMPRedirectCount ()
+  public long getSMPRedirectCount ()
   {
     return 0;
   }
